test(tui): split plugin catalog tests
